pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentmesh"
version = "0.1.0"
description = "AgentMesh Project - Knowledge Base Adapters"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "asyncpg>=0.28.0",
    "neo4j>=5.14.0",
    "mcp>=1.0.0",
    "aiosqlite>=0.19.0",
    "aiohttp>=3.9.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langgraph>=0.6.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "crewai>=0.200.0",
    "openai>=1.100.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["adapters", "mcp_server", "services", "dummy_agents"]

[project.scripts]
agentmesh-mcp = "mcp_server.server:main"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "types-pyyaml>=6.0.12.20250915",
]
